+++ main.go
@@ -2,9 +2,11 @@ package main
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"os"
 	"path"
+	"path/filepath"
 	"runtime"
 	"strings"
 	"time"
@@ -14,6 +16,12 @@ import (
 	"github.com/mlange-42/beecs/params"
 	"github.com/spf13/cobra"
 	"golang.org/x/exp/rand"
+)
+
+const (
+	PARAMETERS = "parameters.json"
+	OBSERVERS  = "observers.json"
+	EXPERIMENT = "experiment.json"
 )
 
 func main() {
@@ -118,15 +126,16 @@ func RootCommand() *cobra.Command {
 
 	root.Flags().StringVarP(&dir, "directory", "d", ".", "Working directory")
 	root.Flags().StringVarP(&outDir, "output", "", "", "Output directory if different from working directory")
-	root.Flags().StringSliceVarP(&paramFiles, "parameters", "p", []string{"parameters.json"}, "Parameter files, processed in the given order")
+	root.Flags().StringSliceVarP(&paramFiles, "parameters", "p", []string{PARAMETERS}, "Parameter files, processed in the given order")
 	root.Flags().StringVarP(&expFile, "experiment", "e", "", "Experiment file for parameter variation")
-	root.Flags().StringVarP(&obsFile, "observers", "o", "observers.json", "Observers file")
+	root.Flags().StringVarP(&obsFile, "observers", "o", OBSERVERS, "Observers file")
 	root.Flags().Float64VarP(&speed, "speed", "s", 0, "Speed limit in ticks per second. Default: 0 (unlimited)")
 	root.Flags().IntVarP(&threads, "threads", "t", runtime.NumCPU(), "Number of threads")
 	root.Flags().IntVarP(&runs, "runs", "r", 1, "Runs per parameter set")
 	root.Flags().IntVarP(&seed, "seed", "", 0, "Super random seed for seed generation. Default: 0 (unseeded)")
 	root.Flags().StringSliceVarP(&overwrite, "overwrite", "x", []string{}, "Overwrite variables like key1=value1,key2=value2")
 
+	root.AddCommand(InitCommand())
 	root.AddCommand(ParametersCommand())
 
 	return root
@@ -168,4 +177,106 @@ func ParametersCommand() *cobra.Command {
 	root.Flags().StringSliceVarP(&paramFiles, "parameters", "p", []string{}, "Optional parameter files, processed in the given order")
 
 	return root
+}
+
+func InitCommand() *cobra.Command {
+	var dir string
+
+	root := &cobra.Command{
+		Use:           "init",
+		Short:         "Initialize templates for an experiment.",
+		Long:          `Initialize templates for an experiment.`,
+		SilenceUsage:  true,
+		SilenceErrors: true,
+		PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
+			return nil
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			parFile := path.Join(dir, PARAMETERS)
+			obsFile := path.Join(dir, OBSERVERS)
+			expFile := path.Join(dir, EXPERIMENT)
+
+			if fileExists(parFile) {
+				return fmt.Errorf("parameter file '%s' already exists", parFile)
+			}
+			if fileExists(obsFile) {
+				return fmt.Errorf("observers file '%s' already exists", obsFile)
+			}
+			if fileExists(expFile) {
+				return fmt.Errorf("experiments file '%s' already exists", expFile)
+			}
+
+			for _, f := range []string{parFile, obsFile, expFile} {
+				err := os.MkdirAll(filepath.Dir(f), os.ModePerm)
+				if err != nil {
+					return err
+				}
+			}
+
+			p := params.Default()
+			err := writeJSON(parFile, &p)
+			if err != nil {
+				return err
+			}
+
+			o := util.ObserversDef{
+				Parameters:      "out/parameters.csv",
+				CsvSeparator:    ",",
+				TimeSeriesPlots: []util.TimeSeriesPlotDef{},
+				Tables:          []util.TableDef{},
+			}
+			err = writeJSON(obsFile, &o)
+			if err != nil {
+				return err
+			}
+
+			e := []experiment.ParameterVariation{
+				{
+					Parameter: "params.InitialStores.Honey",
+					SequenceFloatRange: &experiment.SequenceFloatRange{
+						Min:    10,
+						Max:    100,
+						Values: 10,
+					},
+				},
+			}
+			err = writeJSON(expFile, &e)
+			if err != nil {
+				return err
+			}
+
+			fmt.Printf("Successfully initialized experiment template in '%s'\n", dir)
+
+			return nil
+		},
+	}
+	root.Flags().StringVarP(&dir, "directory", "d", ".", "Working directory")
+
+	return root
+}
+
+func writeJSON(path string, value any) error {
+	js, err := json.MarshalIndent(value, "", "    ")
+	if err != nil {
+		return err
+	}
+	f, err := os.Create(path)
+	if err != nil {
+		return err
+	}
+	if _, err = f.Write(js); err != nil {
+		return err
+	}
+	return f.Close()
+}
+
+func fileExists(name string) bool {
+	_, err := os.Stat(name)
+	if err == nil {
+		return true
+	}
+	if errors.Is(err, os.ErrNotExist) {
+		return false
+	}
+	return false
 }